refactor: 관리자페이지 관심사에 따라 컴포넌트 분리
