+++ src/origin/App.tsx
@@ -1,30 +1,9 @@
 import { useState } from 'react';
-import { CartPage } from './components/CartPage.tsx';
-import { AdminPage } from './components/AdminPage.tsx';
-import { Coupon } from '../types.ts';
-
-const initialCoupons: Coupon[] = [
-  {
-    name: '5000원 할인 쿠폰',
-    code: 'AMOUNT5000',
-    discountType: 'amount',
-    discountValue: 5000
-  },
-  {
-    name: '10% 할인 쿠폰',
-    code: 'PERCENT10',
-    discountType: 'percentage',
-    discountValue: 10
-  }
-];
+import { CartPage } from './pages/CartPage.tsx';
+import { AdminPage } from './pages/AdminPage.tsx';
 
 const App = () => {
-  const [coupons, setCoupons] = useState<Coupon[]>(initialCoupons);
   const [isAdmin, setIsAdmin] = useState(false);
-
-  const handleCouponAdd = (newCoupon: Coupon) => {
-    setCoupons((prevCoupons) => [...prevCoupons, newCoupon]);
-  };
 
   return (
     <div className="min-h-screen bg-gray-100">
